refactor: move the `config` module to the CLI
The config file will be used by the CLI only, not by the YARA-X library as a whole. Every configurable setting in the library is set through some method in `Compiler`, `Scanner` or any of the types exported by the library.
